pyproject.toml
[tool.poetry]
name = "test-task-sdsolution"
version = "0.1.0"
description = ""
authors = ["mmeerrccyy <taras.mercury@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.112.1"}
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
